requirements.txt
django==4.1
djangorestframework
markdown